pyproject.toml
[project]
name = "nuts"
version = "0.1.0"
description = ""
authors = [
    {name = "sasha",email = "sashavipro@gmail.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "wagtail (>=7.2.1,<8.0.0)",
    "wagtailmedia (>=0.17.2,<0.18.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-unfold (>=0.78.1,<0.79.0)",
    "django-countries (>=8.2.0,<9.0.0)",
    "django-cities-light (>=3.10.2,<4.0.0)",
    "django-select2 (>=8.4.8,<9.0.0)",
    "django-ninja (>=1.5.3,<2.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.1.1,<8.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "django-modeltranslation (>=0.19.19,<0.20.0)",
    "django-vite (>=3.1.0,<4.0.0)",
    "gunicorn (>=25.1.0,<26.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.0,<0.16.0)",
    "pre-commit (>=4.5.1,<5.0.0)"
]

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

# ==== Ruff ====
[tool.ruff]
exclude = [
    ".bzr", ".direnv", ".eggs", ".git", ".git-rewrite", ".hg", ".mypy_cache",
    ".nox", ".pants.d", ".pytype", ".ruff_cache", ".svn", ".tox", ".venv",
    "__pypackages__", "_build", "buck-out", "build", "dist", "node_modules",
    "venv", "*/migrations/*.py", "staticfiles/*", "manage.py", "settings.py",
    "Config/*",
]
line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = [
  "F", "E", "W", "C90", "I", "N", "UP", "YTT", "ASYNC", "D", "S", "BLE",
  "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "DJ", "EM", "EXE", "FA", "ISC",
  "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF",
  "SLOT", "SIM", "TID", "TCH", "INT", "PTH", "ERA", "PD", "PGH", "PL",
  "TRY", "FLY", "PERF", "RUF"
]
ignore = [
  "S101", "RUF012", "SIM102", "UP038", "D203", "D213", "COM812", "ISC001",
  "RUF001", "RUF002", "RUF003", "ERA001"
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=mysite.settings.base --reuse-db --import-mode=importlib"
python_files = [
    "tests.py",
    "test_*.py",
]

# ==== Coverage ====
[tool.coverage.run]
source = ["."]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "manage.py",
    "mysite/settings/*",
    "mysite/wsgi.py",
    "mysite/celery.py"
]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
]

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "mysite.settings.base"
